# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lee Pump App API swagger documentation",
        "description": "This is a sample API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/upload/upload-file": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "upload file function ",
                "description": "upload file function",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "description": "file",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/bo/login": {
            "get": {
                "tags": [
                    "BO"
                ],
                "summary": "",
                "description": "login to Binary Option",
                "operationId": "loginBO",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bo/boInfo": {
            "get": {
                "tags": [
                    "BO"
                ],
                "summary": "",
                "description": "boInfo",
                "operationId": "boInfo",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bo/deposit": {
            "post": {
                "tags": [
                    "BO"
                ],
                "summary": "",
                "description": "bo deposit",
                "operationId": "bodeposit",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "amount to deposit",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/news/news-list": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "",
                "description": "News",
                "operationId": "newsList",
                "parameters": [
                    {
                        "name": "language",
                        "in": "query",
                        "description": "en/cn/in/vn/th",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "news_type",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/news/banner-list": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "",
                "description": "News",
                "operationId": "bannerList",
                "parameters": [
                    {
                        "name": "language",
                        "in": "query",
                        "description": "en/cn",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}